requirements.txt
dgl
torch
numpy
pandas
scikit-learn
pyyaml
kaggle
